+++ src/ldclient_config.erl
@@ -41,6 +41,12 @@
     redis_prefix => string(),
     cache_ttl => integer(), % Any negative integer is parsed as an infinite TTL, zero is parsed as testing mode
     use_ldd => boolean(),
+    send_events => boolean(),
+    file_datasource => boolean(),
+    file_paths => [string()],
+    file_auto_update => boolean(),
+    file_poll_interval => pos_integer(),
+    file_allow_duplicate_keys => boolean(),
     ssl_options => [ssl:tls_option()]
 }.
 % Settings stored for each running SDK instance
@@ -64,7 +70,7 @@
 -define(DEFAULT_POLLING_UPDATE_REQUESTOR, ldclient_update_requestor_httpc).
 -define(MINIMUM_POLLING_INTERVAL, 30).
 -define(USER_AGENT, "ErlangClient").
--define(VERSION, "1.1.2").
+-define(VERSION, "1.2.0").
 -define(EVENT_SCHEMA, "3").
 -define(DEFAULT_OFFLINE, false).
 -define(DEFAULT_REDIS_HOST, "127.0.0.1").
@@ -74,6 +80,12 @@
 -define(DEFAULT_REDIS_PREFIX, "launchdarkly").
 -define(DEFAULT_CACHE_TTL, 15).
 -define(DEFAULT_USE_LDD, false).
+-define(DEFAULT_SEND_EVENTS, true).
+-define(DEFAULT_FILE_DATASOURCE, false).
+-define(DEFAULT_FILE_PATHS, []).
+-define(DEFAULT_FILE_AUTO_UPDATE, false).
+-define(DEFAULT_FILE_POLL_INTERVAL, 1000).
+-define(DEFAULT_FILE_ALLOW_DUPLICATE_KEYS, false).
 -define(DEFAULT_SSL_OPTIONS, []).
 
 %%===================================================================
@@ -119,6 +131,12 @@ parse_options(SdkKey, Options) when is_list(SdkKey), is_map(Options) ->
     RedisPassword = maps:get(redis_password, Options, ?DEFAULT_REDIS_PASSWORD),
     RedisPrefix = maps:get(redis_prefix, Options, ?DEFAULT_REDIS_PREFIX),
     CacheTtl = maps:get(cache_ttl, Options, ?DEFAULT_CACHE_TTL),
+    SendEvents = maps:get(send_events, Options, ?DEFAULT_SEND_EVENTS),
+    FileDataSource = maps:get(file_datasource, Options, ?DEFAULT_FILE_DATASOURCE),
+    FilePaths = maps:get(file_paths, Options, ?DEFAULT_FILE_PATHS),
+    FileAutoUpdate = maps:get(file_auto_update, Options, ?DEFAULT_FILE_AUTO_UPDATE),
+    FilePollInterval = maps:get(file_poll_interval, Options, ?DEFAULT_FILE_POLL_INTERVAL),
+    FileAllowDuplicateKeys = maps:get(file_allow_duplicate_keys, Options, ?DEFAULT_FILE_ALLOW_DUPLICATE_KEYS),
     SslOptions = maps:get(ssl_options, Options, ?DEFAULT_SSL_OPTIONS),
     #{
         sdk_key => SdkKey,
@@ -143,6 +161,12 @@ parse_options(SdkKey, Options) when is_list(SdkKey), is_map(Options) ->
         redis_prefix => RedisPrefix,
         cache_ttl => CacheTtl,
         use_ldd => UseLdd,
+        send_events => SendEvents,
+        file_datasource => FileDataSource,
+        file_paths => FilePaths,
+        file_auto_update => FileAutoUpdate,
+        file_poll_interval => FilePollInterval,
+        file_allow_duplicate_keys => FileAllowDuplicateKeys,
         ssl_options => SslOptions
     }.