+++ test/ldclient_eval_SUITE.erl
@@ -57,7 +57,14 @@
     variation_out_of_range/1,
     extra_fields/1,
     missing_some_fields/1,
-    missing_all_fields/1
+    missing_all_fields/1,
+    missing_rollout_for_rule/1,
+    missing_rollout_for_rule_match_rule/1,
+    fallthrough_no_rollout_or_variation/1,
+    fallthrough_rollout_in_experiment/1,
+    fallthrough_rollout_not_in_experiment/1,
+    rule_match_rollout_in_experiment/1,
+    rule_match_rollout_not_in_experiment/1
 ]).
 
 %%====================================================================
@@ -112,7 +119,14 @@ all() ->
         variation_out_of_range,
         extra_fields,
         missing_some_fields,
-        missing_all_fields
+        missing_all_fields,
+        missing_rollout_for_rule,
+        missing_rollout_for_rule_match_rule,
+        fallthrough_no_rollout_or_variation,
+        fallthrough_rollout_in_experiment,
+        fallthrough_rollout_not_in_experiment,
+        rule_match_rollout_in_experiment,
+        rule_match_rollout_not_in_experiment
     ].
 
 init_per_suite(Config) ->
@@ -153,6 +167,9 @@ create_flags() ->
     ok = ldclient_update_stream_server:process_event(#{event => <<"put">>, data => PutData2}, ldclient_storage_ets, another1).
 
 extract_events(Events) ->
+    extract_events(Events, false).
+
+extract_events(Events, false) ->
     [
         {Key, Type, Variation, VariationValue, DefaultValue, Reason, PrereqOf} ||
         #{
@@ -164,6 +181,23 @@ extract_events(Events) ->
                 default := DefaultValue,
                 eval_reason := Reason,
                 prereq_of := PrereqOf
+            }
+        } <- Events
+    ];
+extract_events(Events, true) ->
+    [
+        {Key, Type, Variation, VariationValue, DefaultValue, Reason, PrereqOf, TrackEvents, IncludeReason} ||
+        #{
+            type := Type,
+            data := #{
+                key := Key,
+                variation := Variation,
+                value := VariationValue,
+                default := DefaultValue,
+                eval_reason := Reason,
+                prereq_of := PrereqOf,
+                trackEvents := TrackEvents,
+                include_reason := IncludeReason
             }
         } <- Events
     ].
@@ -706,4 +740,63 @@ missing_all_fields(_) ->
         {<<"missing-all-fields">>, feature_request, null, null, "foo", ExpectedReason, null}
     ]),
     ActualEvents = lists:sort(extract_events(Events)),
+    ExpectedEvents = ActualEvents.
+
+missing_rollout_for_rule(_) -> 
+    {{2,<<"FallthroughValue">>,fallthrough}, Events} = ldclient_eval:flag_key_for_user(default, <<"missing-rollout-for-rule">>, #{key => <<"user123">>}, "DefaultValue"),
+    ExpectedEvents = lists:sort([{<<"missing-rollout-for-rule">>, feature_request, 2, <<"FallthroughValue">>, "DefaultValue", fallthrough, null}]),
+    ActualEvents = lists:sort(extract_events(Events)),
+    ExpectedEvents = ActualEvents.
+
+missing_rollout_for_rule_match_rule(_) ->
+    % For this test the user key matters because the rule matches keys containing "maybe".
+    {{null, "DefaultValue", {error, malformed_flag}}, Events} = ldclient_eval:flag_key_for_user(default, <<"missing-rollout-for-rule">>, #{key => <<"key1Maybe">>}, "DefaultValue"),
+    ExpectedEvents = lists:sort([{<<"missing-rollout-for-rule">>, feature_request, null, "DefaultValue", "DefaultValue", {error, malformed_flag}, null}]),
+    ActualEvents = lists:sort(extract_events(Events)),
+    ExpectedEvents = ActualEvents.
+
+fallthrough_no_rollout_or_variation(_) ->
+    {{null, "DefaultValue", {error, malformed_flag}}, Events} = ldclient_eval:flag_key_for_user(default, <<"fallthrough-no-rollout-or-variation">>, #{key => <<"user123">>}, "DefaultValue"),
+    ExpectedEvents = lists:sort([{<<"fallthrough-no-rollout-or-variation">>, feature_request, null, "DefaultValue", "DefaultValue", {error, malformed_flag}, null}]),
+    ActualEvents = lists:sort(extract_events(Events)),
+    ExpectedEvents = ActualEvents.
+
+fallthrough_rollout_in_experiment(_) ->
+    ExpectedReason = {fallthrough, in_experiment},
+    {{0, <<"a">>, ExpectedReason}, Events} =
+        ldclient_eval:flag_key_for_user(default, <<"experiment-traffic-allocation-v2">>, #{key => <<"userKeyA">>}, "foo"),
+    ExpectedEvents = lists:sort([
+        {<<"experiment-traffic-allocation-v2">>, feature_request, 0, <<"a">>, "foo", ExpectedReason, null, true, true}
+    ]),
+    ActualEvents = lists:sort(extract_events(Events, true)),
+    ExpectedEvents = ActualEvents.
+
+fallthrough_rollout_not_in_experiment(_) ->
+    ExpectedReason = fallthrough,
+    {{1, <<"b">>, ExpectedReason}, Events} =
+        ldclient_eval:flag_key_for_user(default, <<"experiment-traffic-allocation-v2">>, #{key => <<"userKeyB">>}, "foo"),
+    ExpectedEvents = lists:sort([
+        {<<"experiment-traffic-allocation-v2">>, feature_request, 1, <<"b">>, "foo", ExpectedReason, null, false, false}
+    ]),
+    ActualEvents = lists:sort(extract_events(Events, true)),
+    ExpectedEvents = ActualEvents.
+
+rule_match_rollout_in_experiment(_) ->
+    ExpectedReason = {rule_match, 0, <<"ab4a9fb3-7e85-429f-8078-23aa70094540">>, in_experiment},
+    {{0, <<"a">>, ExpectedReason}, Events} =
+        ldclient_eval:flag_key_for_user(default, <<"experiment-traffic-allocation-v2-rules">>, #{key => <<"userKeyA">>}, "foo"),
+    ExpectedEvents = lists:sort([
+        {<<"experiment-traffic-allocation-v2-rules">>, feature_request, 0, <<"a">>, "foo", ExpectedReason, null, true, true}
+    ]),
+    ActualEvents = lists:sort(extract_events(Events, true)),
+    ExpectedEvents = ActualEvents.
+
+rule_match_rollout_not_in_experiment(_) ->
+    ExpectedReason = {rule_match, 0, <<"ab4a9fb3-7e85-429f-8078-23aa70094540">>},
+    {{1, <<"b">>, ExpectedReason}, Events} =
+        ldclient_eval:flag_key_for_user(default, <<"experiment-traffic-allocation-v2-rules">>, #{key => <<"userKeyB">>}, "foo"),
+    ExpectedEvents = lists:sort([
+        {<<"experiment-traffic-allocation-v2-rules">>, feature_request, 1, <<"b">>, "foo", ExpectedReason, null, true, false}
+    ]),
+    ActualEvents = lists:sort(extract_events(Events, true)),
     ExpectedEvents = ActualEvents.